backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.3.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.1.0
fastapi==0.104.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
idna==3.10
mangum==0.19.0
motor==3.3.2
passlib==1.7.4
Pillow==10.1.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.5.0
pydantic_core==2.14.1
pymongo==4.6.0
python-decouple==3.8
python-dotenv==1.0.0
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.2
requests==2.32.4
rsa==4.9.1
sendgrid==6.10.0
six==1.17.0
sniffio==1.3.1
starkbank-ecdsa==2.2.0
starlette==0.27.0
stripe==8.5.0
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
ruff==0.5.5
pytest==8.3.2
pytest-asyncio==0.23.7
